+++ scripts/support/runner/__init__.py
@@ -1,0 +1,22 @@
+# Copyright (c) 2017 Linaro Limited.
+#
+# SPDX-License-Identifier: Apache-2.0
+
+from .core import ZephyrBinaryRunner
+
+# We import these here to ensure the ZephyrBinaryRunner subclasses are
+# defined; otherwise, ZephyrBinaryRunner.create_for_shell_script()
+# won't work.
+
+# Explicitly silence the unused import warning.
+# flake8: noqa: F401
+from . import arc
+from . import bossac
+from . import dfu
+from . import esp32
+from . import nios2
+from . import nrfjprog
+from . import openocd
+from . import pyocd
+
+__all__ = ['ZephyrBinaryRunner']
